enable warnings and fix most of it
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@6196 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ source/de/anomic/kelondro/text/citationPrototype/CitationReferenceRow.java
@@ -91,7 +91,7 @@ public final class CitationReferenceRow implements Reference /*, Cloneable*/ {
         this.entry.setCol(col_lother, lother);
         this.entry.setCol(col_urlLength, urlLength);
         this.entry.setCol(col_urlComps, urlComps);
-        this.entry.setCol(col_typeofurl, new byte[]{(byte) typeofurl});
+        this.entry.setCol(col_typeofurl, new byte[]{typeofurl});
         this.entry.setCol(col_reserve, 0);
     }
